Criação do hook useRoom
