+++ src/pages/Room.tsx
@@ -4,6 +4,7 @@ import { useParams } from 'react-router-dom'
 import logoImg from '../assets/images/logo.svg'
 
 import { useAuth } from '../hooks/useAuth'
+import { useRoom } from '../hooks/useRoom'
 
 import { database } from '../services/firebase'
 
@@ -12,27 +13,6 @@ import { RoomCode } from '../components/RoomCode'
 
 import '../styles/room.scss'
 import { Question } from '../components/Question'
-
-type FirebaseQuestions = Record<string, {
-  author: {
-    name: string;
-    avatar: string;
-  };
-  content: string;
-  isAnswered: boolean;
-  isHighlighted: boolean;
-}>
-
-type QuestionType = {
-  id: string;
-  author: {
-    name: string;
-    avatar: string;
-  };
-  content: string;
-  isAnswered: boolean;
-  isHighlighted: boolean;
-}
 
 type RoomParams = {
   id: string;
@@ -45,30 +25,11 @@ export function Room() {
   const roomId = params.id;
 
   const [newQuestion, setNewQuestion] = useState('')
-  const [questions, setQuestions] = useState<QuestionType[]>([])
-  const [title, setTitle] = useState('')
 
-  useEffect(() => {
-    const roomRef = database.ref(`rooms/${roomId}`)
-
-    roomRef.on('value', room => {
-      const databaseRoom = room.val();
-      const firebaseQuestions: FirebaseQuestions = databaseRoom.questions ?? {};
-      
-      const parsedQuestions = Object.entries(firebaseQuestions).map(([key, value]) => {
-        return {
-          id: key,
-          content: value.content,
-          author: value.author,
-          isHighlighted: value.isHighlighted,
-          isAnswered: value.isAnswered,
-        }
-      })
-
-      setTitle(databaseRoom.title)
-      setQuestions(parsedQuestions)
-    })
-  }, [roomId])
+  const {
+    title,
+    questions,
+  } = useRoom(roomId)
   
   async function handleSendQuestion(e: FormEvent) {
     e.preventDefault()
